+++ packages/ui-prompting/stories/Input.story.tsx
@@ -1,9 +1,10 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Questions } from '../src/components';
 import type { PromptQuestion } from '../src';
 import { initIcons } from '@sap-ux/ui-components';
-import { useI18nBundle, useStorage } from './utils';
+import { useStorage } from './utils';
 import { TRANSLATE_EVENT_UPDATE } from '../src/context/TranslationContext';
+import type { I18nBundle, TranslationEntry } from '@sap-ux/ui-components';
 
 export default { title: 'Basic/Input' };
 
@@ -58,6 +59,51 @@ const questions: PromptQuestion[] = [
         }
     }
 ];
+
+interface CustomTranslationEntry extends TranslationEntry {
+    dummyPath: string;
+}
+
+const I18N_BUNDLE_KEY = 'ui-prompting-i18n-bundle';
+const getI18nBundle = (): I18nBundle<CustomTranslationEntry> => {
+    let i18nBundle: I18nBundle<CustomTranslationEntry> = {};
+    try {
+        i18nBundle = JSON.parse(
+            window.localStorage.getItem(I18N_BUNDLE_KEY) ?? ''
+        ) as I18nBundle<CustomTranslationEntry>;
+    } catch (e) {
+        i18nBundle = {};
+    }
+    return i18nBundle;
+};
+
+const updateI18nBundle = (i18nBundle: I18nBundle): void => {
+    window.localStorage.setItem(I18N_BUNDLE_KEY, JSON.stringify(i18nBundle));
+};
+
+export function useI18nBundle(): [I18nBundle, (question: string, entry: TranslationEntry) => void, string[]] {
+    const [i18nBundle, setI18nBundle] = useState(getI18nBundle());
+    const [pendingQuestions, setPendingQuestions] = useState<string[]>([]);
+    const updateBundle = (question: string, entry: TranslationEntry) => {
+        // Simulate backend call - call with timout
+        pendingQuestions.push(question);
+        setPendingQuestions([...pendingQuestions]);
+        setTimeout(() => {
+            if (!i18nBundle[entry.key.value]) {
+                i18nBundle[entry.key.value] = [{ ...entry, dummyPath: 'test' }];
+                updateI18nBundle(i18nBundle);
+                setI18nBundle({ ...i18nBundle });
+                // Update pending questions
+                const index = pendingQuestions.indexOf(question);
+                if (index !== -1) {
+                    pendingQuestions.splice(index, 1);
+                    setPendingQuestions([...pendingQuestions]);
+                }
+            }
+        }, 1000);
+    };
+    return [i18nBundle, updateBundle, pendingQuestions];
+}
 
 export const input = (): JSX.Element => {
     const [saveValues] = useStorage();
